Refactroring as needed.
