+++ src/test/java/resources/testng.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<suite name="Login and Home" verbose="1" >
+    <test name="Regression suite" >
+        <classes>
+            <class name="com.generic.functional.automation.ui.tests.login.LoginLogoutAutomation"/>
+            <class name="com.generic.functional.automation.ui.tests.login.HomeScreenAutomation"/>
+        </classes>
+    </test>
+</suite>
